+++ permission/src/commonMain/kotlin/tech/kotlinlang/permission/result/RecordAudioPermissionResult.kt
@@ -1,0 +1,7 @@
+package tech.kotlinlang.permission.result
+
+sealed interface RecordAudioPermissionResult {
+    data object Granted: RecordAudioPermissionResult
+    data object NotAllowed: RecordAudioPermissionResult
+    data object Denied: RecordAudioPermissionResult
+}